+++ CHANGELOG.md
@@ -1,3 +1,31 @@
+## Version 1.23.1
+> 10 July, 2024
+
+- Internal improvements and bug fixes.
+
+## Version 1.23
+> 10 July, 2024
+
+- Improves design of feeds
+- Improves following / followers sorting
+
+## Version 1.22.2
+> 10 July, 2024
+
+- Fixes on avatar upload
+- Fixes on posts
+
+## Version 1.22.1
+> 9 July, 2024
+
+- Adds notifications for commenents
+- Fixes on comments
+
+## Version 1.22.0
+> 9 July, 2024
+
+- Adds comments
+
 ## Version 1.21.1
 > 6 July, 2024